Moved existing tests to separate Integration namespace
Restrucuted the tests so that separate unit tests (non-integration) can
be created, each following uniform directory structures.
