+++ tests/Mongofill/Tests/Integration/MongoDBTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mongofill\Tests;
+namespace Mongofill\Tests\Integration;
 
 use MongoCode;
