+++ tests/Mongofill/Tests/Integration/MongoGridFSTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Mongofill\Tests;
+namespace Mongofill\Tests\Integration;
 
 use MongoGridFS;
 
 class MongoGridFSTest extends TestCase
 {
-    const EXAMPLE_BIN_FILE = '/../../resources/MongoGridFS/example.png';
+    const EXAMPLE_BIN_FILE = '/../../../resources/MongoGridFS/example.png';
 
     public function setUp()
     {
