+++ tests/Mongofill/Tests/Integration/MongoTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mongofill\Tests;
+namespace Mongofill\Tests\Integration;
 
 use Mongo;
 use MongoClient;
